Establish a 'communication protocol' between logic and GUI. That protocol includes the most minimal interaction between Logic and GUI (but let be them almost independent of each other, which is very crucial!).
Logic provides only three methods, which can be called by the GUI:
(1) 'get_hardware_constraints' passes the hardware constraints directly to the GUI.
In order to be consistent here about the parameter set, a distinct set of those is defined in the pulser_interface. Only these should be used.
(2) 'get_func_config' retrieves all functions and their definitions and parameter sets. Based on this list the ViewWidget will be set.
(3) 'get_param_config' passes the additional parameters in the same manner as 'get_func_config'.

GUI provides in essence 3 necessary parameters to the Logic:
(1) 'get_current_channels' tells, which of the channel configuration the user has chosen.
(2) 'get_current_function_list'  tells which Functions the user wants to use in the GUI. (Based on this the parameters are chosen in the GUI)
(3) 'get_current_table_config' right now it is a list with the column headers, indicating the structure of the table.

GUI will now react on every change in the func_config of the Logic. Logic must establish the pulse creation and sampling with the given basic functions. Although GUI is now working in this infrastructure, a lot of functionality is still missing. Also a lot of internal restructure of the GUI will follow.