Changed the buttons of the upload section of the pulsed_maingui. You can now manually sample/write the Block_Ensemble, upload the file to the pulse generator and load the uploaded asset into the pulser channel.
Hardware files changed to support the above functionality.

Also changed the sequence_generator_logic to only have a single method to sample ensembles and write them to files. Two file writing modes are supported. One is memory efficient but takes a lot more time. The other one consumes a huge amount of memory for larger Waveforms but is very fast.

The Waveform class is now useless.