Implement a most possible general alignment state machine like structure for general magnet alignment. Now any types of alignment Fluorescence, ODMR, SSR, ... are possible in 1D 2D and even 3D. There is a possibility of making continuous scans while the magnet is moving (mostly for fluorescence alignment needed.) A very easy continue procedure and a highly customizable pathway of the magnet is now possible. This logic can handle in principle an arbitrary amount of magnet axes.
Currently, a proper implementation was done for 2D Fluorescence alignment, where a snakewise pathway was used.
Add also a (general) saving procedure.
Create some signals to which a GUI can attach to, for monitoring changes of the logic.
Tasks are not fully used right now, but they will be implemented on the way properly at some point.

All methods lack a lot of documentation, which will follow in the review process of that module.