Adjusted the former pulsed_measurement_logic parameter "aom_delay_s" in a more general way. Renamed it to "laser_trigger_delay_s" and moved the GUI input widget into the analysis settings dialogue (away from pulse extraction). Implemented proper initialization and status-saving of this parameter.
Also removed anything related to "sample_mode" from the pulser hardware modules since this does not need to be checked by the hardware anymore. The hardware modules will simply search for suitable filenames and do not need to be aware of what is possible in other hardware formats.

Removed the WFMX_header.py module from the repository since the header creation is now included in samples_write_methods.py.