Fixed a bug with the new scientific spinboxes in the pulsed GUI analysis settings.
Removed the setting of CW mode in the ODMR logic after the measurement has ended. This was causing the microwave to output shortly a single frequency CW but with the power specified in CW mode and not the one used for the sweep. This could cause serious harm if this power is much higher than the one used in the sweep.

Also removed the call of the current fit function after the measurement has ended in ODMR logic. Depending on what is set as fit function in the combo box this was leading to errors.