Converted the pulse block editor and the block organizer editor into classes found in pulse_editors.py. Reintroduced the editors in the maingui_new and they seem to work properly.
Completely changed how the PulseBlockElements, PulseBlocks, PulseBlockEnsembles and PulseSequences are represented. The length is now specified in seconds which is more general than bins. Carefully adjusted sampling methods and other stuff to account for those changes. Also removed the underscores of these object classes. They are now in CamelCaseNotation.

Basic methods are not yet reworked and will not work currently.

Old serialized pulse objects will not work either but it should not throw an error if they are still present... they just get ignored.