Changed the pulse analysis and extraction logic from a linear master-slave chain to be directly used by the pulsed_measurement_logic. Stripped the pulse_extraction_logic from fastcounter hardware access. This will hopefully increase stability because now only one logic module is accessing the hardware module during measurement.
Introduced the possibility to store raw count data with a name tag in the pulsed_measurement_logic to resume the measurement later on. The old raw data will then be added to the new one. Works only in logic (tested with dummy) and can not be used via GUI yet.