Fixed a bug when asking the AWG70k for the currently loaded asset.
Introduced a new functionality to pulsed measurements. Now one can pass a tag string when stopping a pulsed measurement. If the string is not empty (default) the measurement raw data is stashed in a dictionary (in pulsed_measurement_logic) with the tag string serving as key. Raw data is only temporarily stashed that way so upon closing Qudi the raw data is lost.
You can then also pass a tag string when starting a pulsed measurement to recall old raw data with the same tag and add it to the newly recorded raw data during this measurement.
This enables the user to resume a measurement while being able to perform arbitrary steps in between.