fix non-QObject class having a Slot
Apparently, in some pyqt5 installations (debian native python/Qt5/pyqt5),
you cannot send a signal to a method decorated with @Slot() if the class
is not derived from QObject.

In a C++ application, this behaviour would be obviously correct as the mechanism
for signal reception/slot declaration is simply missing.

PyQt4 allows this "impossible" thing and for PyQt5 it only fails in a fraction
of environments. Anaconda Qt5/PyQt5 seems generally fine, Debian Qt5/PyQt5 is
usually not.