Increased robustness of GUI during activation when strange parameters are passed from logic.
added a channel config to the pulser dummy.

Rewritten write_to_file method for .wfm waveform files (AWG 7k and 5k series, compatible in 70k as well). Now the writing to file is equally performing for .wfmx and .wfm. The only difference is that .wfm is using a bit more memory (approx. factor 1.5) but this comes from the fact that analog and digital samples have to be written together and not sequentially.
Tested on AWG 70k.