Tested the whole direct write toolchain with the dummy modules. Works now after a couple of bug-fixes for both Ensembles and Sequences.
Changed the Sample_sequence method in sequence_generator_logic quite drastically. Now the naming convention for the waveforms is the sequence name plus the number of the current sequence step the ensemble is representing if rotating wave should be preserved across the entire sequence. In case the rotating wave should not be preserved (or only internally in every ensemble) the naming of the waveforms is the same as for just sampling a single Ensemble (the ensemble name plus channel descriptor).