In predefined_methods added a new function called _get_multiple_mw_element which allows for the generation of a simple microwave element with double or triple sine. It could be generalized to replace the  function _get_mw_element.
Some small changes to magnet_logic like increasing the number of decimal points on the ROI.
Adjusting the Box size of the ROI with the selected region in the 2D scan.
Hope to have fixed the issue of the ROI not moving correctly with the
 position the user moved to.

 Added some caller tags to the poi_manager as it was always giving an error message when it got an unknown caller tag ( in our case from the magnet_logic ).
Hope that is okay.

Some adjustment to the trace_analysis logic because of recent changes to fit_logic.