fixed the issues described by Christoph:
There is now an option letting you select, which count trace should be displayed on top. Additionally you can select which traces are to be plotted.

However, the save still saves all the traces.

----
Also fixed a small error in returning the analogue values for multiple channels.

Thanks to Ben for the Tip!