Implemented a new dock Widget in the cavity stabilisation gui with the control parameters for the cavity stabilisation. You can enter now the threshold, the resolution (voltage adjustment steps) and the average number directly in the gui, and you can choose the cavity mode (reflection or transmission) via checkbox.
The shown scan numbers as well as the points per scan in the plots can now be changed via config option. The default values are 5 for the shown scan numbers and 1000 for the points per scan.

Further if you enter a scan resolution that is higher than the maximum resolution of your scanning device there will be a warning and the scan resolution will be set to the maximum value.

You dont need to write a maximum clock frequency in the config. This option is not needed any more.

A bug has been fixed when entering -1V as start value for the scan. this fix has to be double checked!