fixed poi-manager: a feature that was more a bug
The PoiManagerLogic was listening for a signal, that the confocal position was changed and then deleting the active poi (if it wasn't done by the optimizer.

This let to the effect, that you could simply loose your POI by changing the focus between windows.

As far as I know, there wasn't any usecase, where you needed to forget your active POI when changing the position, so this should not hurt anyone.