make_sampling_functions_externalizeable: This is a small feature update.
It searches the additionally files imported by the SequenceGeneratorLogic for the variables _math_func and func_config and appends them to the class-functions.
The effect is, that now also sampling functions can be externalized as was already done for the predefined methods.

Also the predefined ranges were extended for more general use.