Added 3D steps scan options to qt GUI.
This is meant to be able to do a 3rd axis scan for every point in the 3d axis. the speed and amount of points can be chosen in the GUI. The 3rd axis to be scanned can not be chosen yet but is decided by the config.
Additionally added check box to choose to only scan one direction and not both to increase scan speed.