Added new option to step scan:
3D step scan.
At each step an additional scan is added along a  third axis, this is so far an analog axis ( _ao_channel ).
For this the way data was sorted was optimized and the saving changed.
There is a new option for saving (_save_3D_measurement ) and the program now saves the data for each line during the measurement as a npy file.
This is meant to reduce data in the ram (as the 3D data is discarded after saving) as well as make sure that data is saved even if the program crashes as 3D scans on average take considerably more time than 2D scans and are therefore more prone to other errors.
Added additional potion "_fast_scan" which will only scan one direction (forward) for each line reducing the measurement time. The stepper still moves back on the same line (not change on the slow axis) but with considerably more speed.