Updated stabilization part of cavity_stabilisation_logic. It now takes multiple values for each power feedback via analogue input to sample out noise using a highest frequency possible.
For this the NIDAQ and the analogue_reader_interface had to be updated.
Also slightly reorganized NIDAQ to account for methods that are needed by multiple users and are intrinsic to this specific hardware but not necessarily all hardwares which can be used with this code (using config to specify value).

This change increases accuracy in the stabilization  (averaging out errors) as well as increasing the speed for this accuracy (no extra software time delay). It comes at the cost of having to restart the stabilization if a change in the averaging number is wanted by the user as the hardware has to be reinitialized.