Added additional functionality:
Additional to measuring the 3D dimension while stepping (partly synchronized) this can now measure the finesse or something similar for eg. a cavity for every point in the image.
For this it steps and then waits until the 3d axis was measured completely and independently and then steps to the next point.
It has been tested on the attocube setup Niv. 3 and seems to work fine.
However it does not have a GUI connection yet and no self sufficient saving capability.