Sometimes the stepper  did not actually move because it got stuck in one axis.
If the stepper has position feedback this needs to be recognized by the program and handled accordingly.
This is implemented by _check_movement_during_measurement and _compare_measured_positions_stepper methods.
Another problem of this sort is handled by changing the position to which the stepper returns at end of each line being the start position of the first line instead of the current line. This  way the probability of the stepper moving out of the desired scan area is reduced.
The optimization had problems either with the stepper getting stuck or making the stepper get stuck by too many repetitions. It now assumes that a certain amount of steps always leads to a movement.
Furthermore the optimization now finds the optimal amount of steps to use on its own. reducing the overall optimization time.
Methods using optimization are adjusted accordingly.