Further step in the process of multiple ranges feature.
Replacing the old functionality to set the range.
Connecting signals to the boxes. Constructing swept frequencies
from the additional spin boxes.
Handling of different methods in the microwave hardware.

Now I am not sure how to continue as I encountered a problem with
connecting the signals. Basically I connect a lot of signals
to the same function and that seems to cause a crash.