First Debug stage of scanner_3D_logic. It is possible to load logic without errors.
Added function in national_instruments_x_series inside the newer analogue scanner function to now also scan positions.
For this a new variable _a_o_pos_ranges had to be introduced. This is a quick fix, but until NIDAQ is rewritten it doesnt make sense to update as it would mean doubling of config info.