Ironing out lots of bugs and inconsistencies in the dummy file.
Decided on making a category based system for the acquisition modes. For now there
is 'Single Scan' and 'Series' category the acquisition modes have to fall in.
These modes sould later on be supported by the logic.

For this also added the function get_available_acquisition_modes into the interface.