Cleaned up and modernized ruamel.yaml dump/load handling to make use of the object oriented API. Could still be improved by enhancing readability of dumped files but this is a pain due to the lacking documentation of ruamel.yaml.
Improved Configuration object API to be useful for editing and creating qudi configurations (not just loading existing ones). Also introduced a lot of constraints and sanity checks on config data in order to provide a robust config editor object that incorporates all requirements for a qudi config in one place.